chore: Remove venv volume bind
The venv volume bind was there to simplify development since it is
harder to inspect file hierarchies that exist within the container
whereas it is much easier to be able to inspect file trees that are
easily inspectable/accessible from a host.

This configuration may "trick" users into believing that their local
venv state is comparable to that of the container had it been built from
scratch. To remove the chance for this confusion, we are now requiring
users to just rebuild the container images from scratch for every venv
change.

Thanks to @oz123 for the guidance here.